# src/main/resources/openapi/api.yaml
openapi: 3.1.0
info:
  title: Highload api
  description: Highload api
  version: 1.0.0

servers:
  - url: 'http://localhost:8080'

paths:
  /api/user/register:
    post:
      summary: RegisterUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'schemas/users.yaml#/components/schemas/UserDTO'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: 'common/response.yaml#/components/schemas/Default200Response'
  /api/group/create:
    post:
      summary: CreateGroup
      parameters:
        - in: header
          name: CurrentUser
          required: true
          schema:
            type: string
          description: Id текущего юзера
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'schemas/group.yaml#/components/schemas/GroupDTO'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: 'common/response.yaml#/components/schemas/Default200Response'
  /api/group/addUser:
    post:
      summary: Add User to group
      parameters:
        - in: header
          name: CurrentUser
          required: true
          schema:
            type: string
          description: Id текущего юзера
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - group
                - username
                - role
              properties:
                group:
                  $ref: 'schemas/group.yaml#/components/schemas/GroupDTO'
                username:
                  type: string
                  description: Имя пользователя для добавления
                role:
                  type: string
                  enum: ["Admin", "Participant"]
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: 'common/response.yaml#/components/schemas/Default200Response'

  /api/ticket/create:
    summary: Create new Ticket
    post:
      parameters:
        - in: header
          name: CurrentUser
          required: true
          schema:
            type: string
          description: Id текущего юзера
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - group
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 100
                description:
                  type: string
                  minLength: 1
                group:
                  $ref: 'schemas/group.yaml#/components/schemas/GroupDTO'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: 'common/response.yaml#/components/schemas/Default200Response'